Reorganise modules, and reduce surface area of public API
