+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "headless_chrome"
-version = "0.1.1"
+version = "0.1.2"
 authors = ["Alistair Roche <alistair@sunburnt.country>"]
 edition = "2018"
 description = "Control Chrome programatically"
