+++ src/lib.rs
@@ -3,21 +3,10 @@
 extern crate log;
 extern crate termcolor;
 
-pub mod browser;
+mod browser;
 pub mod cdtp;
-pub mod element;
-pub mod helpers;
-pub mod keys;
-pub mod logging;
-pub mod point;
-pub mod process;
-pub mod tab;
-pub mod transport;
-pub mod waiting_call_registry;
-pub mod web_socket_connection;
 
-pub use browser::{Browser, LaunchOptions};
-pub use tab::Tab;
+pub use browser::{Browser, LaunchOptions, Tab};
 
 #[cfg(feature = "nightly")]
 #[doc(include = "../README.md")]
