+++ tests/logging.rs
